// frontend/src/external/client/api/orval/generated/zod/company-scouts/company-scouts.zod.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * inselfy API
 * OpenAPI spec version: 1.0.0
 */
import * as zod from 'zod';

import { zodParams } from '../../../zod-params';

/**
 * スカウト送信（企業認証）
 * @summary Send a scout message
 */
export const companyScoutsSendScoutBodyCandidateIdOneRegExp = new RegExp('^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$');
export const companyScoutsSendScoutBodySubjectMax = 200;

export const companyScoutsSendScoutBodyBodyMax = 5000;



export const CompanyScoutsSendScoutBody = zod.object({
  "candidateId": zod.string(zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["candidateId"],"validator":"string"})).regex(companyScoutsSendScoutBodyCandidateIdOneRegExp, zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["candidateId"],"validator":"regex"})).describe('UUID 形式の識別子').describe('候補者ユーザーID'),
  "jobPostingId": zod.string(zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["jobPostingId"],"validator":"string"})).nullish().describe('関連求人ID'),
  "templateId": zod.string(zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["templateId"],"validator":"string"})).nullish().describe('使用テンプレートID'),
  "subject": zod.string(zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["subject"],"validator":"string"})).max(companyScoutsSendScoutBodySubjectMax, zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["subject"],"validator":"max"})).describe('件名'),
  "body": zod.string(zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["body"],"validator":"string"})).max(companyScoutsSendScoutBodyBodyMax, zodParams({"operationId":"CompanyScouts_sendScout","location":"body","schemaName":"CompanyScoutsSendScoutBody","fieldPath":["body"],"validator":"max"})).describe('本文')
}).describe('スカウト送信リクエスト')

/**
 * スカウトへ返信（企業認証）
 * @summary Reply to a scout as the company
 */
export const companyScoutsCompanyScoutReplyBodyBodyMax = 5000;



export const CompanyScoutsCompanyScoutReplyBody = zod.object({
  "body": zod.string(zodParams({"operationId":"CompanyScouts_companyScoutReply","location":"body","schemaName":"CompanyScoutsCompanyScoutReplyBody","fieldPath":["body"],"validator":"string"})).max(companyScoutsCompanyScoutReplyBodyBodyMax, zodParams({"operationId":"CompanyScouts_companyScoutReply","location":"body","schemaName":"CompanyScoutsCompanyScoutReplyBody","fieldPath":["body"],"validator":"max"})).describe('本文')
}).describe('企業からのスカウト返信リクエスト')